# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.3.0
  description: API segura do Família Segura
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /catalog/apps:
    get:
      operationId: listCatalogApps
      tags: [rules]
      responses:
        "200":
          description: Catálogo de apps conhecidos (globais + descobertos pela família)
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/CatalogApp" }
        "401": { $ref: "#/components/responses/Unauthorized" }
...
  /family/children/{childId}/apps:
    post:
      operationId: createAppRule
      tags: [rules]
      parameters:
        - $ref: "#/components/parameters/ChildId"
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/AppRuleInput" }
      responses:
        "201":
          description: Regra criada
          content:
            application/json:
              schema: { $ref: "#/components/schemas/AppRule" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "404": { $ref: "#/components/responses/NotFound" }
        "409": { $ref: "#/components/responses/Conflict" }
  /family/children/{childId}/apps/{appId}:
    delete:
      operationId: deleteAppRule
      tags: [rules]
      parameters:
        - $ref: "#/components/parameters/ChildId"
        - $ref: "#/components/parameters/AppId"
      responses:
        "204":
          description: Regra removida
        "401": { $ref: "#/components/responses/Unauthorized" }
        "404": { $ref: "#/components/responses/NotFound" }
  /family/children/{childId}/routines:
    post:
      operationId: createRoutine
      tags: [rules]
      parameters:
        - $ref: "#/components/parameters/ChildId"
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/RoutineInput" }
      responses:
        "201":
          description: Rotina criada
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Routine" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "404": { $ref: "#/components/responses/NotFound" }
  /family/children/{childId}/grants:
    post:
      operationId: createTimeGrant
      tags: [rules]
      parameters:
        - $ref: "#/components/parameters/ChildId"
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/TimeGrantInput" }
      responses:
        "201":
          description: Tempo extra liberado somente para hoje
          content:
            application/json:
              schema: { $ref: "#/components/schemas/TimeGrant" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "404": { $ref: "#/components/responses/NotFound" }
...
  /family/apps/{appId}/rules:
    patch:
      operationId: updateAppRule
      tags: [rules]
      parameters:
        - $ref: "#/components/parameters/AppId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AppRuleUpdate"
      responses:
        "200":
          description: Rule updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AppRule"
        "401": { $ref: "#/components/responses/Unauthorized" }
        "404": { $ref: "#/components/responses/NotFound" }
  /family/routines/{routineId}:
    patch:
      operationId: updateRoutine
      tags: [rules]
      parameters:
        - $ref: "#/components/parameters/RoutineId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RoutineUpdate"
      responses:
        "200":
          description: Routine updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Routine"
        "401": { $ref: "#/components/responses/Unauthorized" }
        "404": { $ref: "#/components/responses/NotFound" }
    delete:
      operationId: deleteRoutine
      tags: [rules]
      parameters:
        - $ref: "#/components/parameters/RoutineId"
      responses:
        "204":
          description: Rotina removida
        "401": { $ref: "#/components/responses/Unauthorized" }
        "404": { $ref: "#/components/responses/NotFound" }
  /family/time-requests:
    post:
      operationId: createTimeRequest
      tags: [requests]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TimeRequestInput"
      responses:
        "201":
          description: Request created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TimeRequest"
        "401": { $ref: "#/components/responses/Unauthorized" }
  /family/time-requests/{requestId}:
    patch:
      operationId: resolveTimeRequest
      tags: [requests]
      parameters:
        - $ref: "#/components/parameters/RequestId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TimeRequestResolution"
      responses:
        "200":
          description: Request resolved
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TimeRequest"
        "401": { $ref: "#/components/responses/Unauthorized" }
        "404": { $ref: "#/components/responses/NotFound" }
...
components:
  parameters:
...
    ChildId:
      in: path
      name: childId
      required: true
      schema: { type: string, format: uuid }
...
    AppId:
      in: path
      name: appId
      required: true
      schema: { type: string, minLength: 1, maxLength: 200 }
...
    RoutineId:
      in: path
      name: routineId
      required: true
      schema: { type: string, format: uuid }
    RequestId:
      in: path
      name: requestId
      required: true
      schema: { type: string, format: uuid }
  responses:
    Unauthorized:
      description: Authentication required
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ErrorResponse" }
    BadRequest:
      description: Invalid request
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ErrorResponse" }
...
    NotFound:
      description: Not found
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ErrorResponse" }
    Conflict:
      description: Conflito (limite atingido ou registro duplicado)
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ErrorResponse" }
...
  schemas:
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error: { type: string }
        code: { type: string }
...
    AppRuleInput:
      type: object
      properties:
        catalogAppId: { type: string, minLength: 1, maxLength: 200 }
        name: { type: string, minLength: 1, maxLength: 80 }
        category: { type: string, minLength: 1, maxLength: 40 }
        androidPackages:
          type: array
          maxItems: 10
          items: { type: string, minLength: 3, maxLength: 200 }
        dailyLimitMinutes: { type: integer, minimum: 0, maximum: 1440 }
        status: { type: string, enum: [allowed, attention, blocked] }
    AppRuleUpdate:
      type: object
      required: [childId]
      properties:
        childId: { type: string, format: uuid }
        status: { type: string, enum: [allowed, attention, blocked] }
        dailyLimitMinutes: { type: integer, minimum: 0, maximum: 1440 }
    RoutineInput:
      type: object
      required: [title, days, startTime, endTime]
      properties:
        title: { type: string, minLength: 1, maxLength: 80 }
        description: { type: string, maxLength: 160 }
        days: { type: string, pattern: "^(dom|seg|ter|qua|qui|sex|sab)(,(dom|seg|ter|qua|qui|sex|sab))*$" }
        startTime: { type: string, pattern: "^([01][0-9]|2[0-3]):[0-5][0-9]$" }
        endTime: { type: string, pattern: "^([01][0-9]|2[0-3]):[0-5][0-9]$" }
        icon: { type: string, maxLength: 40 }
        enabled: { type: boolean }
    RoutineUpdate:
      type: object
      properties:
        title: { type: string, minLength: 1, maxLength: 80 }
        description: { type: string, maxLength: 160 }
        days: { type: string, pattern: "^(dom|seg|ter|qua|qui|sex|sab)(,(dom|seg|ter|qua|qui|sex|sab))*$" }
        startTime: { type: string, pattern: "^([01][0-9]|2[0-3]):[0-5][0-9]$" }
        endTime: { type: string, pattern: "^([01][0-9]|2[0-3]):[0-5][0-9]$" }
        icon: { type: string, maxLength: 40 }
        enabled: { type: boolean }
    TimeGrantInput:
      type: object
      required: [appId, minutes]
      properties:
        appId: { type: string, minLength: 1, maxLength: 200 }
        minutes: { type: integer, minimum: 5, maximum: 240 }
    TimeGrant:
      type: object
      required: [id, childId, appId, minutes, validOn, source, createdAt]
      properties:
        id: { type: string, format: uuid }
        childId: { type: string, format: uuid }
        appId: { type: string }
        minutes: { type: integer }
        validOn: { type: string }
        source: { type: string, enum: [manual, request] }
        createdAt: { type: string, format: date-time }
    TimeRequestInput:
      type: object
      required: [childId, appId, requestedMinutes, message]
      properties:
        childId: { type: string, format: uuid }
        appId: { type: string, minLength: 1, maxLength: 200 }
        requestedMinutes: { type: integer, minimum: 5, maximum: 120 }
        message: { type: string, maxLength: 240 }
    TimeRequestResolution:
      type: object
      required: [status]
      properties:
        status: { type: string, enum: [approved, denied] }
        grantedMinutes: { type: integer, minimum: 5, maximum: 240 }
...
    CatalogApp:
      type: object
      required: [id, name, category, icon, iconColor, androidPackages]
      properties:
        id: { type: string }
        name: { type: string }
        category: { type: string }
        icon: { type: string }
        iconColor: { type: string }
        androidPackages:
          type: array
          items: { type: string }
...
    AppRule:
      type: object
      required: [id, childId, appId, appName, category, icon, iconColor, androidPackages, usageTodayMinutes, dailyLimitMinutes, extraTodayMinutes, effectiveLimitMinutes, status]
      properties:
        id: { type: string, format: uuid }
        childId: { type: string, format: uuid }
        appId: { type: string }
        appName: { type: string }
        category: { type: string }
        icon: { type: string }
        iconColor: { type: string }
        androidPackages:
          type: array
          items: { type: string }
        usageTodayMinutes: { type: integer }
        dailyLimitMinutes:
          type: integer
          description: No painel do responsável é o limite base. Na visão da criança já inclui o tempo extra de hoje.
        extraTodayMinutes: { type: integer }
        effectiveLimitMinutes: { type: integer }
        status: { type: string, enum: [allowed, attention, blocked] }
    Routine:
      type: object
      required: [id, childId, title, description, days, startTime, endTime, enabled, icon]
      properties:
        id: { type: string, format: uuid }
        childId: { type: string, format: uuid }
        title: { type: string }
        description: { type: string }
        days: { type: string }
        startTime: { type: string }
        endTime: { type: string }
        enabled: { type: boolean }
        icon: { type: string }
    TimeRequest:
      type: object
      required: [id, childId, childName, appId, appName, requestedMinutes, message, status, createdAt]
      properties:
        id: { type: string, format: uuid }
        childId: { type: string, format: uuid }
        childName: { type: string }
        appId: { type: string }
        appName: { type: string }
        requestedMinutes: { type: integer }
        message: { type: string }
        status: { type: string, enum: [pending, approved, denied] }
        createdAt: { type: string, format: date-time }
        resolvedAt: { type: [string, "null"], format: date-time }
...