# lib/api-zod/src/generated/types/catalogApp.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API segura do Família Segura
 * OpenAPI spec version: 0.3.0
 */

export interface CatalogApp {
  id: string;
  name: string;
  category: string;
  icon: string;
  iconColor: string;
  androidPackages: string[];
}